[Refactor] Use MMCV MODEL_REGISTRY (#495)
* remove registry.py in datasets/

* remove registry.py in ops/pointnet_modules/

* remove registry.py in models/

* refactor builder using mmcv.MODELS registry

* update mmcv min version requirement

* change min mmcv version requirement in mmdet3d/__init__.py

* min mmcv version to 1.3.2
