--- mmdet3d/datasets/registry.py
@@ -1,3 +1,0 @@
-from mmcv.utils import Registry
-
-OBJECTSAMPLERS = Registry('Object sampler')
