+++ mmdet3d/models/__init__.py
@@ -1,5 +1,6 @@
 from .backbones import *  # noqa: F401,F403
-from .builder import (build_backbone, build_detector, build_fusion_layer,
+from .builder import (FUSION_LAYERS, MIDDLE_ENCODERS, VOXEL_ENCODERS,
+                      build_backbone, build_detector, build_fusion_layer,
                       build_head, build_loss, build_middle_encoder, build_neck,
                       build_roi_extractor, build_shared_head,
                       build_voxel_encoder)
@@ -10,7 +11,6 @@ from .losses import *  # noqa: F401,F403
 from .middle_encoders import *  # noqa: F401,F403
 from .model_utils import *  # noqa: F401,F403
 from .necks import *  # noqa: F401,F403
-from .registry import FUSION_LAYERS, MIDDLE_ENCODERS, VOXEL_ENCODERS
 from .roi_heads import *  # noqa: F401,F403
 from .voxel_encoders import *  # noqa: F401,F403
