+++ mmdet3d/models/fusion_layers/vote_fusion.py
@@ -2,7 +2,7 @@ import torch
 from torch import nn as nn
 
 from mmdet3d.core.bbox import Coord3DMode, points_cam2img
-from ..registry import FUSION_LAYERS
+from ..builder import FUSION_LAYERS
 from . import apply_3d_transformation, bbox_2d_transform, coord_2d_transform
 
 EPS = 1e-6
