+++ mmdet3d/models/middle_encoders/pillar_scatter.py
@@ -2,7 +2,7 @@ import torch
 from mmcv.runner import auto_fp16
 from torch import nn
 
-from ..registry import MIDDLE_ENCODERS
+from ..builder import MIDDLE_ENCODERS
 
 
 @MIDDLE_ENCODERS.register_module()
