--- mmdet3d/models/registry.py
@@ -1,5 +1,0 @@
-from mmcv.utils import Registry
-
-VOXEL_ENCODERS = Registry('voxel_encoder')
-MIDDLE_ENCODERS = Registry('middle_encoder')
-FUSION_LAYERS = Registry('fusion_layer')
