+++ mmdet3d/models/voxel_encoders/pillar_encoder.py
@@ -4,7 +4,7 @@ from mmcv.runner import force_fp32
 from torch import nn
 
 from mmdet3d.ops import DynamicScatter
-from ..registry import VOXEL_ENCODERS
+from ..builder import VOXEL_ENCODERS
 from .utils import PFNLayer, get_paddings_indicator
