+++ mmdet3d/models/voxel_encoders/voxel_encoder.py
@@ -5,7 +5,7 @@ from torch import nn
 
 from mmdet3d.ops import DynamicScatter
 from .. import builder
-from ..registry import VOXEL_ENCODERS
+from ..builder import VOXEL_ENCODERS
 from .utils import VFELayer, get_paddings_indicator
