+++ mmdet3d/ops/pointnet_modules/builder.py
@@ -1,4 +1,6 @@
-from .registry import SA_MODULES
+from mmcv.utils import Registry
+
+SA_MODULES = Registry('point_sa_module')
 
 
 def build_sa_module(cfg, *args, **kwargs):
